requirements.txt
streamlit
pandas
geopandas
fiona
pyogrio
requests
pillow
numpy
sounddevice